src: introduce compatibility functions
Due to upstream changes the way we call the functions has changed
slightly, instead of rewriting all of it now. I am introducing some
convenience functions to map the structures properly. This is because of
time constraint, but I will refactor it at a later stage.

These functions are also in place to avoid breaking code already relying
on the old structure.

Special thanks to @SebastianStehle for helping with debugging and fixing
these.

Signed-off-by: Alexander Alemayhu <alexander@alemayhu.com>